% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greet.R
\name{greet}
\alias{greet}
\title{Kindly greet the first user of your package}
\usage{
greet(whom)
}
\arguments{
\item{whom}{\link{character(1)}\cr The person/entity that shall be greeted.}
}
\value{
Character string that greets.
}
\description{
Kindly greet the first user of your package
}
\examples{
# greet the world ...
greet(whom = "World")

# ... or a range of numbers
greet(c(1, 2, 3, 4))
}
