% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plt_DW}
\alias{plt_DW}
\title{Shoot root_DW}
\format{
A data frame with 3 variables and 22 observations:
\describe{
\item{\code{SE}}{mycorrhiza and herbivore}
\item{\code{Shoot}}{Shoot dry weight}
\item{\code{Root}}{Root dry weight}
}
}
\usage{
plt_DW
}
\description{
A dataset of plant dry weight during tripartie interaction
}
\keyword{datasets}
